Break modal component
This is to demo what Happo can help you discover automatically. In this
case I manually introduced a change that will render the modal
invisible. In the real world, these types of mistakes are made on
accident and are generally hard to catch while doing manual testing or
unit testing.